requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langchain-huggingface
chromadb
sentence-transformers
youtube-transcript-api
python-dotenv